+++ RealmServer/Handlers/OnJoinChannel.cs
@@ -7,7 +7,7 @@ namespace RealmServer.Handlers
     {
         public static void Handler(RealmServerSession session, CMSG_JOIN_CHANNEL handler)
         {
-            Console.WriteLine(handler.Channel);
+            
         }
     }
 }